src/requirements.txt
bleak              0.22.3
Flask              3.1.0
python-dotenv      1.0.1
picamera2